Move TestUndoMountSnapFailsInCopyData, TestSideInfo* and TestGadgetDefaults* tests to snapstate_install_test.go (thanks pedronis).
