+++ overlord/snapstate/snapstate_test.go
@@ -8331,94 +8331,6 @@ func (s *snapmgrTestSuite) TestDisableDoesNotEnableAgain(c *C) {
 	c.Assert(ts, IsNil)
 }
 
-func (s *snapmgrTestSuite) TestUndoMountSnapFailsInCopyData(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	chg := s.state.NewChange("install", "install a snap")
-	opts := &snapstate.RevisionOptions{Channel: "some-channel"}
-	ts, err := snapstate.Install(context.Background(), s.state, "some-snap", opts, s.user.ID, snapstate.Flags{})
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.fakeBackend.copySnapDataFailTrigger = filepath.Join(dirs.SnapMountDir, "some-snap/11")
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	expected := fakeOps{
-		{
-			op:     "storesvc-snap-action",
-			userID: 1,
-		},
-		{
-			op: "storesvc-snap-action:action",
-			action: store.SnapAction{
-				Action:       "install",
-				InstanceName: "some-snap",
-				Channel:      "some-channel",
-			},
-			revno:  snap.R(11),
-			userID: 1,
-		},
-		{
-			op:   "storesvc-download",
-			name: "some-snap",
-		},
-		{
-			op:    "validate-snap:Doing",
-			name:  "some-snap",
-			revno: snap.R(11),
-		},
-		{
-			op:  "current",
-			old: "<no-current>",
-		},
-		{
-			op:   "open-snap-file",
-			path: filepath.Join(dirs.SnapBlobDir, "some-snap_11.snap"),
-			sinfo: snap.SideInfo{
-				RealName: "some-snap",
-				SnapID:   "some-snap-id",
-				Channel:  "some-channel",
-				Revision: snap.R(11),
-			},
-		},
-		{
-			op:    "setup-snap",
-			name:  "some-snap",
-			path:  filepath.Join(dirs.SnapBlobDir, "some-snap_11.snap"),
-			revno: snap.R(11),
-		},
-		{
-			op:   "copy-data.failed",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/11"),
-			old:  "<no-old>",
-		},
-		{
-			op:   "remove-snap-data-dir",
-			name: "some-snap",
-			path: filepath.Join(dirs.SnapDataDir, "some-snap"),
-		},
-		{
-			op:    "undo-setup-snap",
-			name:  "some-snap",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap/11"),
-			stype: "app",
-		},
-		{
-			op:   "remove-snap-dir",
-			name: "some-snap",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap"),
-		},
-	}
-	// start with an easier-to-read error if this fails:
-	c.Assert(s.fakeBackend.ops.Ops(), DeepEquals, expected.Ops())
-	c.Assert(s.fakeBackend.ops, DeepEquals, expected)
-}
-
 func (s *snapmgrTestSuite) TestRefreshFailureCausesErrorReport(c *C) {
 	var errSnap, errMsg, errSig string
 	var errExtra map[string]string
@@ -10811,128 +10723,6 @@ defaults:
 	c.Assert(defls, DeepEquals, map[string]interface{}{"foo": "bar"})
 }
 
-func (s *snapmgrTestSuite) TestGadgetDefaultsAreNormalizedForConfigHook(c *C) {
-	var mockGadgetSnapYaml = `
-name: canonical-pc
-type: gadget
-`
-	var mockGadgetYaml = []byte(`
-defaults:
-  otheridididididididididididididi:
-    foo:
-      bar: baz
-      num: 1.305
-
-volumes:
-    volume-id:
-        bootloader: grub
-`)
-
-	info := snaptest.MockSnap(c, mockGadgetSnapYaml, &snap.SideInfo{Revision: snap.R(2)})
-	err := ioutil.WriteFile(filepath.Join(info.MountDir(), "meta", "gadget.yaml"), mockGadgetYaml, 0644)
-	c.Assert(err, IsNil)
-
-	gi, err := gadget.ReadInfo(info.MountDir(), nil)
-	c.Assert(err, IsNil)
-	c.Assert(gi, NotNil)
-
-	snapName := "some-snap"
-	hooksup := &hookstate.HookSetup{
-		Snap:        snapName,
-		Hook:        "configure",
-		Optional:    true,
-		IgnoreError: false,
-		TrackError:  false,
-	}
-
-	var contextData map[string]interface{}
-	contextData = map[string]interface{}{"patch": gi.Defaults}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-	c.Assert(hookstate.HookTask(s.state, "", hooksup, contextData), NotNil)
-}
-
-func makeInstalledMockCoreSnap(c *C) {
-	coreSnapYaml := `name: core
-version: 1.0
-type: os
-`
-	snaptest.MockSnap(c, coreSnapYaml, &snap.SideInfo{
-		RealName: "core",
-		Revision: snap.R(1),
-	})
-}
-
-func (s *snapmgrTestSuite) TestGadgetDefaults(c *C) {
-	r := release.MockOnClassic(false)
-	defer r()
-
-	makeInstalledMockCoreSnap(c)
-
-	// using MockSnap, we want to read the bits on disk
-	snapstate.MockSnapReadInfo(snap.ReadInfo)
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	s.prepareGadget(c)
-
-	snapPath := makeTestSnap(c, "name: some-snap\nversion: 1.0")
-
-	ts, _, err := snapstate.InstallPath(s.state, &snap.SideInfo{RealName: "some-snap", SnapID: "some-snap-id", Revision: snap.R(1)}, snapPath, "", "edge", snapstate.Flags{})
-	c.Assert(err, IsNil)
-
-	var m map[string]interface{}
-	runHooks := tasksWithKind(ts, "run-hook")
-
-	c.Assert(taskKinds(runHooks), DeepEquals, []string{
-		"run-hook[install]",
-		"run-hook[configure]",
-		"run-hook[check-health]",
-	})
-	err = runHooks[1].Get("hook-context", &m)
-	c.Assert(err, IsNil)
-	c.Assert(m, DeepEquals, map[string]interface{}{"use-defaults": true})
-}
-
-func (s *snapmgrTestSuite) TestGadgetDefaultsNotForOS(c *C) {
-	r := release.MockOnClassic(false)
-	defer r()
-
-	// using MockSnap, we want to read the bits on disk
-	snapstate.MockSnapReadInfo(snap.ReadInfo)
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "core", nil)
-
-	s.prepareGadget(c)
-
-	const coreSnapYaml = `
-name: core
-type: os
-version: 1.0
-`
-	snapPath := makeTestSnap(c, coreSnapYaml)
-
-	ts, _, err := snapstate.InstallPath(s.state, &snap.SideInfo{RealName: "core", SnapID: "core-id", Revision: snap.R(1)}, snapPath, "", "edge", snapstate.Flags{})
-	c.Assert(err, IsNil)
-
-	var m map[string]interface{}
-	runHooks := tasksWithKind(ts, "run-hook")
-
-	c.Assert(taskKinds(runHooks), DeepEquals, []string{
-		"run-hook[install]",
-		"run-hook[configure]",
-		"run-hook[check-health]",
-	})
-	// use-defaults flag is part of hook-context which isn't set
-	err = runHooks[1].Get("hook-context", &m)
-	c.Assert(err, Equals, state.ErrNoState)
-}
-
 func (s *snapmgrTestSuite) TestNoReRefreshInUpdate(c *C) {
 	s.state.Lock()
 	defer s.state.Unlock()
@@ -10959,37 +10749,6 @@ func (s *snapmgrTestSuite) TestNoReRefreshInUpdate(c *C) {
 	// NoReRefresh is consumed and consulted when creating the taskset
 	// but is not copied into SnapSetup
 	c.Check(snapsup.Flags.NoReRefresh, Equals, false)
-}
-
-func (s *snapmgrTestSuite) TestGadgetDefaultsInstalled(c *C) {
-	makeInstalledMockCoreSnap(c)
-
-	// using MockSnap, we want to read the bits on disk
-	snapstate.MockSnapReadInfo(snap.ReadInfo)
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	s.prepareGadget(c)
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{{RealName: "some-snap", SnapID: "some-snap-id", Revision: snap.R(1)}},
-		Current:  snap.R(1),
-		SnapType: "app",
-	})
-
-	snapPath := makeTestSnap(c, "name: some-snap\nversion: 1.0")
-
-	ts, _, err := snapstate.InstallPath(s.state, &snap.SideInfo{RealName: "some-snap", SnapID: "some-snap-id", Revision: snap.R(2)}, snapPath, "", "edge", snapstate.Flags{})
-	c.Assert(err, IsNil)
-
-	var m map[string]interface{}
-	runHooks := tasksWithKind(ts, "run-hook")
-
-	c.Assert(runHooks[0].Kind(), Equals, "run-hook")
-	err = runHooks[0].Get("hook-context", &m)
-	c.Assert(err, Equals, state.ErrNoState)
 }
 
 func (s *snapmgrTestSuite) TestTransitionCoreTasksNoUbuntuCore(c *C) {
@@ -12221,52 +11980,6 @@ func (s *snapmgrTestSuite) TestSnapManagerRefreshSchedule(c *C) {
 		c.Check(scheduleStr, Equals, t.out)
 		c.Check(legacy, Equals, false)
 	}
-}
-
-func (s *snapmgrTestSuite) TestSideInfoPaid(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-	opts := &snapstate.RevisionOptions{Channel: "channel-for-paid"}
-	ts, err := snapstate.Install(context.Background(), s.state, "some-snap", opts, s.user.ID, snapstate.Flags{})
-	c.Assert(err, IsNil)
-
-	chg := s.state.NewChange("install", "install paid snap")
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	// verify snap has paid sideinfo
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, IsNil)
-	c.Check(snapst.CurrentSideInfo().Paid, Equals, true)
-	c.Check(snapst.CurrentSideInfo().Private, Equals, false)
-}
-
-func (s *snapmgrTestSuite) TestSideInfoPrivate(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-	opts := &snapstate.RevisionOptions{Channel: "channel-for-private"}
-	ts, err := snapstate.Install(context.Background(), s.state, "some-snap", opts, s.user.ID, snapstate.Flags{})
-	c.Assert(err, IsNil)
-
-	chg := s.state.NewChange("install", "install private snap")
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	// verify snap has private sideinfo
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, IsNil)
-	c.Check(snapst.CurrentSideInfo().Private, Equals, true)
-	c.Check(snapst.CurrentSideInfo().Paid, Equals, false)
 }
 
 func (s *snapmgrTestSuite) TestUpdateLayoutsChecksFeatureFlag(c *C) {
